requirements.txt
Django>=2.2.0,<2.3.0
psycopg2>=2.8.2,<2.9.0
gunicorn>=19.9.0,<19.10.0
djangorestframework>=3.9.0,<3.10.0
django-rest-auth>=0.9.5,<0.10.0
django-allauth>=0.39.1,<0.40.0
flake8>=3.6.0,<3.7.0
Pillow>=5.3.0,<5.4.0
celery>=4.3.0,<4.4.0
redis>=3.2.1,<3.3.0
importlib-metadata==4.13.0